- Removed the 8-character limit on endpic names from the parser. (Though it   might still be present in the texture manager; I don't remember.) - Fixed: EndSequence needs a proper constructor. - Some more GCC warning removals.
SVN r1021 (trunk)

+++ tools/lemon/lemon.c
@@ -1674,9 +1674,9 @@ FILE *err;
   spcnt += k;
   for(; argv[i]; i++) fprintf(err," %s",argv[i]);
   if( spcnt<20 ){
-    fprintf(err,"\n%*s^-- here\n",spcnt,"");
+    fprintf(err,"\n%*s^-- here\n",(int)spcnt,"");
   }else{
-    fprintf(err,"\n%*shere --^\n",spcnt-7,"");
+    fprintf(err,"\n%*shere --^\n",(int)spcnt-7,"");
   }
 }
 
@@ -1915,7 +1915,7 @@ void OptPrint(){
     switch( op[i].type ){
       case OPT_FLAG:
       case OPT_FFLAG:
-        fprintf(errstream,"  -%-*s  %s\n",max,op[i].label,op[i].message);
+        fprintf(errstream,"  -%-*s  %s\n",(int)max,op[i].label,op[i].message);
         break;
       case OPT_INT:
       case OPT_FINT:
